simplify about page
